Rename ident package based on feedback in PR
