+++ parser/schema.go
@@ -12,7 +12,6 @@ import (
 
 	"encr.dev/parser/est"
 	"encr.dev/pkg/errlist"
-	"encr.dev/pkg/identifiers"
 	schema "encr.dev/proto/encore/parser/schema/v1"
 )
 
@@ -123,7 +122,7 @@ func (p *parser) resolveType(pkg *est.Package, file *est.File, expr ast.Expr, ty
 					RawTag:          opts.RawTag,
 				}
 				if f.QueryStringName == "" {
-					f.QueryStringName = identifiers.ConvertIdentifierTo(f.Name, identifiers.SnakeCase)
+					f.QueryStringName = idents.Convert(f.Name, idents.SnakeCase)
 				}
 
 				st.Fields = append(st.Fields, f)
